refactor(service): rename entrypoint file for consistency with other project services codebase
