+++ service/tests/test_routes.py
@@ -1,5 +1,5 @@
 import pytest
-from ..routes import app
+from ..main import app
 import subprocess
 import time
